refactor: StudioImageInput
- preview 그라데이션 추가
- Controls hover시 bg 변경
- components/studio로 이동
