+++ client/src/app/components/studio/StudioAddButton.tsx
