+++ client/src/app/components/studio/StudioBadge.tsx
